+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.17.18
+
+* added parameter immediate to ExponentialBackoff.launch
+* fixed a bug with ExponentialBackoff.launch